refac(EH): Modularize infrastructure
